// src/app/models/paciente.ts
export interface Paciente {
  idUsuario: number;
  rutUsuario: string;
  pnombreUsuario: string;
  snombreUsuario: string;
  apaternoUsuario: string;
  amaternoUsuario: string;
  emailUsuario: string;
  telefonoUsuario: string;
  fechaNacimientoUsuario: string;
  contrasenaUsuario: string;
  activo: boolean;
  fechaCreacionUsuario: string;
  idGenero: number;
  idDireccion: number;
  idRol: number;
}